refactor:(add interface)
