+++ src/main/java/camp/nextstep/edu/kitchenpos/menu/bo/MenuBo.java
@@ -1,12 +1,12 @@
 package camp.nextstep.edu.kitchenpos.menu.bo;
 
 import camp.nextstep.edu.kitchenpos.menu.domain.Menu;
+import camp.nextstep.edu.kitchenpos.menu.domain.MenuDao;
 import camp.nextstep.edu.kitchenpos.menu.domain.MenuProduct;
-import camp.nextstep.edu.kitchenpos.menu.infra.MenuDao;
-import camp.nextstep.edu.kitchenpos.menu.infra.MenuProductDao;
-import camp.nextstep.edu.kitchenpos.menugroup.infra.MenuGroupDao;
+import camp.nextstep.edu.kitchenpos.menu.domain.MenuProductDao;
+import camp.nextstep.edu.kitchenpos.menugroup.domain.MenuGroupDao;
 import camp.nextstep.edu.kitchenpos.product.domain.Product;
-import camp.nextstep.edu.kitchenpos.product.infra.ProductDao;
+import camp.nextstep.edu.kitchenpos.product.domain.ProductDao;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
