+++ src/main/java/camp/nextstep/edu/kitchenpos/menu/domain/MenuDao.java
@@ -1,0 +1,13 @@
+package camp.nextstep.edu.kitchenpos.menu.domain;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface MenuDao {
+
+    Menu save(final Menu menu);
+
+    Optional<Menu> findById(final Long id);
+
+    List<Menu> findAll();
+}
