+++ src/main/java/camp/nextstep/edu/kitchenpos/menu/domain/MenuProductDao.java
@@ -1,0 +1,15 @@
+package camp.nextstep.edu.kitchenpos.menu.domain;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface MenuProductDao {
+
+    MenuProduct save(final MenuProduct entity);
+
+    Optional<MenuProduct> findById(final Long id);
+
+    List<MenuProduct> findAll();
+
+    List<MenuProduct> findAllByMenuId(final Long menuId);
+}
