+++ src/main/java/camp/nextstep/edu/kitchenpos/menugroup/domain/MenuGroupDao.java
@@ -1,0 +1,16 @@
+package camp.nextstep.edu.kitchenpos.menugroup.domain;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface MenuGroupDao {
+
+    MenuGroup save(final MenuGroup entity);
+
+    Optional<MenuGroup> findById(final Long id);
+
+    List<MenuGroup> findAll();
+
+    boolean existsById(final Long id);
+
+}
