+++ src/main/java/camp/nextstep/edu/kitchenpos/order/domain/OrderDao.java
@@ -1,0 +1,16 @@
+package camp.nextstep.edu.kitchenpos.order.domain;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface OrderDao {
+
+    Order save(final Order entity);
+
+    Optional<Order> findById(final Long id);
+
+    List<Order> findAll();
+
+    boolean existsByOrderTableIdAndOrderStatusIn(final Long orderTableId, final List<String> orderStatuses);
+
+}
