+++ src/main/java/camp/nextstep/edu/kitchenpos/product/domain/ProductDao.java
@@ -1,0 +1,14 @@
+package camp.nextstep.edu.kitchenpos.product.domain;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface ProductDao {
+
+    Product save(final Product entity);
+
+    Optional<Product> findById(final Long id);
+
+    List<Product> findAll();
+
+}
