+++ src/main/java/camp/nextstep/edu/kitchenpos/product/infra/ProductRDB.java
@@ -1,6 +1,7 @@
-package camp.nextstep.edu.kitchenpos.menu.infra;
+package camp.nextstep.edu.kitchenpos.product.infra;
 
-import camp.nextstep.edu.kitchenpos.menu.domain.Menu;
+import camp.nextstep.edu.kitchenpos.product.domain.Product;
+import camp.nextstep.edu.kitchenpos.product.domain.ProductDao;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
@@ -15,14 +16,14 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class MenuDao {
-    private static final String TABLE_NAME = "menu";
+public class ProductRDB implements ProductDao {
+    private static final String TABLE_NAME = "product";
     private static final String KEY_COLUMN_NAME = "id";
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
-    public MenuDao(final DataSource dataSource) {
+    public ProductRDB(final DataSource dataSource) {
         jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
         jdbcInsert = new SimpleJdbcInsert(dataSource)
                 .withTableName(TABLE_NAME)
@@ -30,13 +31,15 @@ public class MenuDao {
         ;
     }
 
-    public Menu save(final Menu menu) {
-        final SqlParameterSource parameters = new BeanPropertySqlParameterSource(menu);
+    @Override
+    public Product save(final Product entity) {
+        final SqlParameterSource parameters = new BeanPropertySqlParameterSource(entity);
         final Number key = jdbcInsert.executeAndReturnKey(parameters);
         return select(key.longValue());
     }
 
-    public Optional<Menu> findById(final Long id) {
+    @Override
+    public Optional<Product> findById(final Long id) {
         try {
             return Optional.of(select(id));
         } catch (final EmptyResultDataAccessException e) {
@@ -44,24 +47,24 @@ public class MenuDao {
         }
     }
 
-    public List<Menu> findAll() {
-        final String sql = "SELECT id, name, price, menu_group_id FROM menu ";
+    @Override
+    public List<Product> findAll() {
+        final String sql = "SELECT id, name, price FROM product";
         return jdbcTemplate.query(sql, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private Menu select(final Long id) {
-        final String sql = "SELECT id, name, price, menu_group_id FROM menu WHERE id = (:id)";
+    private Product select(final Long id) {
+        final String sql = "SELECT id, name, price FROM product WHERE id = (:id)";
         final SqlParameterSource parameters = new MapSqlParameterSource()
                 .addValue("id", id);
         return jdbcTemplate.queryForObject(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private Menu toEntity(final ResultSet resultSet) throws SQLException {
-        final Menu entity = new Menu();
-        entity.setId(resultSet.getLong("id"));
+    private Product toEntity(final ResultSet resultSet) throws SQLException {
+        final Product entity = new Product();
+        entity.setId(resultSet.getLong(KEY_COLUMN_NAME));
         entity.setName(resultSet.getString("name"));
         entity.setPrice(resultSet.getBigDecimal("price"));
-        entity.setMenuGroupId(resultSet.getLong("menu_group_id"));
         return entity;
     }
 }
