+++ src/main/java/camp/nextstep/edu/kitchenpos/tablegroup/bo/TableGroupBo.java
@@ -1,11 +1,11 @@
 package camp.nextstep.edu.kitchenpos.tablegroup.bo;
 
+import camp.nextstep.edu.kitchenpos.order.domain.OrderDao;
 import camp.nextstep.edu.kitchenpos.order.domain.OrderStatus;
-import camp.nextstep.edu.kitchenpos.order.infra.OrderDao;
 import camp.nextstep.edu.kitchenpos.ordertable.domain.OrderTable;
-import camp.nextstep.edu.kitchenpos.ordertable.infra.OrderTableDao;
+import camp.nextstep.edu.kitchenpos.ordertable.domain.OrderTableDao;
 import camp.nextstep.edu.kitchenpos.tablegroup.domain.TableGroup;
-import camp.nextstep.edu.kitchenpos.tablegroup.infra.TableGroupDao;
+import camp.nextstep.edu.kitchenpos.tablegroup.domain.TableGroupDao;
 import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.Comparator;
