+++ src/main/java/camp/nextstep/edu/kitchenpos/tablegroup/domain/TableGroupDao.java
@@ -1,0 +1,14 @@
+package camp.nextstep.edu.kitchenpos.tablegroup.domain;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface TableGroupDao {
+
+    TableGroup save(final TableGroup entity);
+
+    Optional<TableGroup> findById(final Long id);
+
+    List<TableGroup> findAll();
+
+}
