+++ src/main/java/camp/nextstep/edu/kitchenpos/tablegroup/infra/TableGroupRDB.java
@@ -1,8 +1,10 @@
-package camp.nextstep.edu.kitchenpos.menu.infra;
+package camp.nextstep.edu.kitchenpos.tablegroup.infra;
 
-import camp.nextstep.edu.kitchenpos.menu.domain.MenuProduct;
+import camp.nextstep.edu.kitchenpos.tablegroup.domain.TableGroup;
+import camp.nextstep.edu.kitchenpos.tablegroup.domain.TableGroupDao;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import javax.sql.DataSource;
@@ -15,14 +17,14 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class MenuProductDao {
-    private static final String TABLE_NAME = "menu_product";
-    private static final String KEY_COLUMN_NAME = "seq";
+public class TableGroupRDB implements TableGroupDao {
+    private static final String TABLE_NAME = "table_group";
+    private static final String KEY_COLUMN_NAME = "id";
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
-    public MenuProductDao(final DataSource dataSource) {
+    public TableGroupRDB(final DataSource dataSource) {
         jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
         jdbcInsert = new SimpleJdbcInsert(dataSource)
                 .withTableName(TABLE_NAME)
@@ -30,13 +32,15 @@ public class MenuProductDao {
         ;
     }
 
-    public MenuProduct save(final MenuProduct entity) {
+    @Override
+    public TableGroup save(final TableGroup entity) {
         final SqlParameterSource parameters = new BeanPropertySqlParameterSource(entity);
         final Number key = jdbcInsert.executeAndReturnKey(parameters);
         return select(key.longValue());
     }
 
-    public Optional<MenuProduct> findById(final Long id) {
+    @Override
+    public Optional<TableGroup> findById(final Long id) {
         try {
             return Optional.of(select(id));
         } catch (final EmptyResultDataAccessException e) {
@@ -44,31 +48,23 @@ public class MenuProductDao {
         }
     }
 
-    public List<MenuProduct> findAll() {
-        final String sql = "SELECT seq, menu_id, product_id, quantity FROM menu_product";
+    @Override
+    public List<TableGroup> findAll() {
+        final String sql = "SELECT id, created_date FROM table_group";
         return jdbcTemplate.query(sql, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    public List<MenuProduct> findAllByMenuId(final Long menuId) {
-        final String sql = "SELECT seq, menu_id, product_id, quantity FROM menu_product WHERE menu_id = (:menuId)";
+    private TableGroup select(final Long id) {
+        final String sql = "SELECT id, created_date FROM table_group WHERE id = (:id)";
         final SqlParameterSource parameters = new MapSqlParameterSource()
-                .addValue("menuId", menuId);
-        return jdbcTemplate.query(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
-    }
-
-    private MenuProduct select(final Long id) {
-        final String sql = "SELECT seq, menu_id, product_id, quantity FROM menu_product WHERE seq = (:seq)";
-        final SqlParameterSource parameters = new MapSqlParameterSource()
-                .addValue("seq", id);
+                .addValue("id", id);
         return jdbcTemplate.queryForObject(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private MenuProduct toEntity(final ResultSet resultSet) throws SQLException {
-        final MenuProduct entity = new MenuProduct();
-        entity.setSeq(resultSet.getLong(KEY_COLUMN_NAME));
-        entity.setMenuId(resultSet.getLong("menu_id"));
-        entity.setProductId(resultSet.getLong("product_id"));
-        entity.setQuantity(resultSet.getLong("quantity"));
+    private TableGroup toEntity(final ResultSet resultSet) throws SQLException {
+        final TableGroup entity = new TableGroup();
+        entity.setId(resultSet.getLong(KEY_COLUMN_NAME));
+        entity.setCreatedDate(resultSet.getObject("created_date", LocalDateTime.class));
         return entity;
     }
 }
