+++ src/test/java/camp/nextstep/edu/kitchenpos/bo/OrderBoTest.java
@@ -8,14 +8,14 @@ import static org.mockito.Mockito.when;
 
 import camp.nextstep.edu.kitchenpos.order.bo.OrderBo;
 import camp.nextstep.edu.kitchenpos.order.domain.Order;
+import camp.nextstep.edu.kitchenpos.order.domain.OrderDao;
 import camp.nextstep.edu.kitchenpos.order.domain.OrderLineItem;
+import camp.nextstep.edu.kitchenpos.order.domain.OrderLineItemDao;
 import camp.nextstep.edu.kitchenpos.order.domain.OrderStatus;
-import camp.nextstep.edu.kitchenpos.order.infra.OrderDao;
-import camp.nextstep.edu.kitchenpos.order.infra.OrderLineItemDao;
 import camp.nextstep.edu.kitchenpos.ordertable.domain.OrderTable;
-import camp.nextstep.edu.kitchenpos.ordertable.infra.OrderTableDao;
+import camp.nextstep.edu.kitchenpos.ordertable.domain.OrderTableDao;
 import camp.nextstep.edu.kitchenpos.tablegroup.domain.TableGroup;
-import camp.nextstep.edu.kitchenpos.tablegroup.infra.TableGroupDao;
+import camp.nextstep.edu.kitchenpos.tablegroup.domain.TableGroupDao;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
