+++ src/test/java/camp/nextstep/edu/kitchenpos/bo/ProductBoTest.java
@@ -7,7 +7,7 @@ import static org.mockito.Mockito.when;
 
 import camp.nextstep.edu.kitchenpos.product.bo.ProductBo;
 import camp.nextstep.edu.kitchenpos.product.domain.Product;
-import camp.nextstep.edu.kitchenpos.product.infra.ProductDao;
+import camp.nextstep.edu.kitchenpos.product.domain.ProductDao;
 import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
